% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ETAD_csv.R
\name{download_ETAD_csv}
\alias{download_ETAD_csv}
\title{Download ETAD data in CSV format}
\usage{
download_ETAD_csv(file)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}
}
\description{
\code{download_ETAD_csv} downloads a comma-separated-values file
of the ETAD dataset.
}
\seealso{
\code{\link{download_ETAD_tsv}}, \code{\link{download_ETAD_dta}},
\code{\link[utils]{write.csv}}
}
