% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowand_ETAD_dta.R
\name{download_ETAD_dta}
\alias{download_ETAD_dta}
\title{Download ETAD data in Stata dta format}
\usage{
download_ETAD_dta(file, version = 14)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}

\item{version}{file version to use. Supports versions 8-15. Default to 14.}
}
\description{
\code{download_ETAD_dta} downloads a Stata dta file of
the ETAD dataset.
}
\seealso{
\code{\link{download_ETAD_csv}}, \code{\link{download_ETAD_tsv}},
\code{\link[haven]{write_dta}}
}
