% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ETAD_tsv.R
\name{download_ETAD_tsv}
\alias{download_ETAD_tsv}
\title{Download ETAD data in TSV format}
\usage{
download_ETAD_tsv(file)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}
}
\description{
\code{download_ETAD_tsv} downloads a tab-separated-values file
of the ETAD dataset.
}
\seealso{
\code{\link{download_ETAD_csv}}, \code{\link{download_ETAD_dta}},
\code{\link[utils]{write.table}}
}
