% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ETAD_csv.R
\name{export_ETAD_csv}
\alias{export_ETAD_csv}
\title{Export ETAD data as a CSV file}
\usage{
export_ETAD_csv(file)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}
}
\description{
\code{export_ETAD_csv} exports a comma-separated-values file
of the ETAD dataset.
}
\seealso{
\code{\link{export_ETAD_tsv}}, \code{\link{export_ETAD_dta}},
\code{\link[utils]{write.csv}}
}
