% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ETAD_tsv.R
\name{export_ETAD_tsv}
\alias{export_ETAD_tsv}
\title{Export ETAD data as a TSV file}
\usage{
export_ETAD_tsv(file)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}
}
\description{
\code{export_ETAD_tsv} exports a tab-separated-values file
of the ETAD dataset.
}
\seealso{
\code{\link{export_ETAD_csv}}, \code{\link{export_ETAD_dta}},
\code{\link[utils]{write.table}}
}
