% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ETAD_dta.R
\name{export_ETAD_dta}
\alias{export_ETAD_dta}
\title{Export ETAD data as a Stata dta file}
\usage{
export_ETAD_dta(file, version = 15)
}
\arguments{
\item{file}{either a character string naming a file or a connection open for
writing.}

\item{version}{file version to use. Supports versions 8-15. Default to 15.}
}
\description{
\code{export_ETAD_dta} exports a Stata dta file of
the ETAD dataset.
}
\seealso{
\code{\link{export_ETAD_csv}}, \code{\link{export_ETAD_tsv}},
\code{\link[haven]{write_dta}}
}
